Save the cart into cookies and rework of the addToCartButton
