+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "eslint": "8.41.0",
     "eslint-config-next": "13.4.3",
+    "js-cookie": "^3.0.5",
     "next": "13.4.3",
     "react": "18.2.0",
     "react-dom": "18.2.0"
