+++ src/components/Header.js
@@ -2,12 +2,18 @@ import Link from "next/link";
 import Logo from "../../public/img/logo.jpg";
 import Image from "next/image";
 import Layout from "./Layout";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import { Store } from "@/utils/Store";
 
 const Header = () => {
 	const { state } = useContext(Store);
 	const { cart } = state;
+	const [cartItemsCount, setCartItemsCount] = useState(0);
+
+	useEffect(() => {
+		setCartItemsCount(cart.cartItems.reduce((a, c) => a + c.quantity, 0));
+	}, [cart.cartItems]);
+
 	return (
 		<header>
 			<Layout>
@@ -26,10 +32,7 @@ const Header = () => {
 				</span>
 				<div>
 					<Link href="/cart">
-						Cart{" "}
-						{cart.cartItems.length > 0 && (
-							<span>{cart.cartItems.reduce((a, c) => a + c.quantity, 0)}</span>
-						)}
+						Cart {cartItemsCount > 0 && <span>{cartItemsCount}</span>}
 					</Link>
 					<Link href="/login">Login</Link>
 				</div>
