+++ src/utils/Store.js
@@ -1,8 +1,11 @@
+import Cookies from "js-cookie";
 import { createContext, useReducer } from "react";
 
 export const Store = createContext();
 
-const initialState = { cart: { cartItems: [] } };
+const initialState = {
+	cart: Cookies.get("cart") ? JSON.parse(Cookies.get("cart")) : { cartItems: [] },
+};
 
 function reducer(state, action) {
 	switch (action.type) {
@@ -14,6 +17,7 @@ function reducer(state, action) {
 						item.name === existItem.name ? newItem : item
 				  )
 				: [...state.cart.cartItems, newItem];
+			Cookies.set("cart", JSON.stringify({ ...state.cart, cartItems }));
 			return { ...state, cart: { ...state.cart, cartItems } };
 		}
 		case "CART_REMOVE_ITEM": {
